refactor: refactor import module
