+++ src/screens/libraryScreens/index.js
@@ -1,4 +1,4 @@
-import LibraryScreen from './LibraryScreen';
+import LibraryScreen from './libraryScreen';
 import ActionLibraryScreen from './actionLibraryScreen';
 import SearchLibraryScreen from './searchLibraryScreen';
