+++ src/screens/screenTopAppBars/index.js
@@ -1,8 +1,8 @@
-import {GamePlayTopAppBar} from '../gameScreens/gamePlayScreen/components';
-import {GameWaitTopAppBar} from '../gameScreens/gameWaitScreen/components';
-import {HomeTopAppBar} from '../homeScreen/components';
-import {LibraryTopAppBar} from '../libraryScreens/LibraryScreen/components';
-import {CreateCardTopAppBar} from '../actionScreens/createCardScreen/components';
+import {GamePlayTopAppBar} from 'src/screens/gameScreens/gamePlayScreen/components';
+import {GameWaitTopAppBar} from 'src/screens/gameScreens/gameWaitScreen/components';
+import {HomeTopAppBar} from 'src/screens/homeScreen/components';
+import {LibraryTopAppBar} from 'src/screens/libraryScreens/libraryScreen/components';
+import {CreateCardTopAppBar} from 'src/screens/actionScreens/createCardScreen/components';
 
 export {
   GamePlayTopAppBar,
